+++ src/main/java/com/soeztrip/travelplanner/service/TripService.java
@@ -4,12 +4,14 @@ package com.soeztrip.travelplanner.service;
 import com.soeztrip.travelplanner.dto.TripDto;
 import com.soeztrip.travelplanner.dto.UserDto;
 import com.soeztrip.travelplanner.model.Trip;
+import com.soeztrip.travelplanner.model.TripRole;
 import com.soeztrip.travelplanner.model.UserEntity;
 import com.soeztrip.travelplanner.model.UserTrip;
 import com.soeztrip.travelplanner.repository.TripRepository;
 import com.soeztrip.travelplanner.repository.UserRepository;
 import com.soeztrip.travelplanner.repository.UserTripRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -25,13 +27,17 @@ public class TripService {
 
     private UserTripRepository userTripRepository;
 
+    private TripRoleService tripRoleService;
+
     @Autowired
     public TripService(TripRepository tripRepository,
                        UserRepository userRepository,
-                       UserTripRepository userTripRepository) {
+                       UserTripRepository userTripRepository,
+                       TripRoleService tripRoleService) {
         this.tripRepository = tripRepository;
         this.userRepository = userRepository;
         this.userTripRepository = userTripRepository;
+        this.tripRoleService = tripRoleService;
     }
 
     public void deleteTrip(Long id) {
@@ -91,6 +97,8 @@ public class TripService {
         UserTrip userTrip = new UserTrip();
         userTrip.setUser(user);
         userTrip.setTrip(trip);
+        TripRole role = tripRoleService.getRoleByName("OWNER");
+        userTrip.setTripRole(role);
         tripRepository.save(trip);
         userTripRepository.save(userTrip);
     }
@@ -99,7 +107,18 @@ public class TripService {
         return tripRepository.existsById(id);
     }
 
+    private String checkUserRole(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String requesterEmail = authentication.getName();
+        return userTripRepository.findRole(id, requesterEmail).orElseThrow();
+    }
+
     public void updateTrip(Long id, TripDto tripDto) {
+        String requesterRole = this.checkUserRole(id);
+        if (!"OWNER".equals(requesterRole) && !"MANAGER".equals(requesterRole)) {
+            throw new RuntimeException("Only the trip owner or manager can modify trip properties");
+        }
+
         Trip trip = this.tripRepository.findById(id).orElseThrow();
         if (tripDto.getStartingDate() != null) {
             trip.setStartingDate(tripDto.getStartingDate());
@@ -118,19 +137,49 @@ public class TripService {
     }
 
     public void addParticipant(Long id, String email) {
+        String requesterRole = checkUserRole(id);
+        if (!"OWNER".equals(requesterRole) && !"MANAGER".equals(requesterRole)) {
+            throw new RuntimeException("Only the trip owner or manager can add participants");
+        }
         UserEntity user = this.userRepository.findByEmail(email).orElseThrow();
         Trip trip = this.tripRepository.findById(id).orElseThrow();
         UserTrip userTrip = new UserTrip();
         userTrip.setUser(user);
         userTrip.setTrip(trip);
+        TripRole role = tripRoleService.getRoleByName("PARTICIPANT");
+        userTrip.setTripRole(role);
         userTripRepository.save(userTrip);
     }
 
     public void removeParticipant(Long id, String email) {
+        String requesterRole = checkUserRole(id);
+        if (!"OWNER".equals(requesterRole) && !"MANAGER".equals(requesterRole)) {
+            throw new RuntimeException("Only the trip owner or manager can remove participants");
+        }
+        UserEntity tripOwner = userTripRepository.findOwner(id).orElseThrow();
+        if (tripOwner.getEmail().equals(email)) {
+            throw new RuntimeException("Owner can not be removed from the trip");
+        }
         UserEntity user = this.userRepository.findByEmail(email).orElseThrow();
         Trip trip = this.tripRepository.findById(id).orElseThrow();
         UserTrip userTrip = userTripRepository.findByUserAndTrip(user, trip);
         userTripRepository.delete(userTrip);
         tripRepository.save(trip);
+    }
+
+    public void changeRole(Long id, String email, String role) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String requesterEmail = authentication.getName();
+        UserEntity tripOwner = userTripRepository.findOwner(id).orElseThrow();
+        if (!tripOwner.getEmail().equals(requesterEmail)) {
+            throw new RuntimeException("Only trip owner can change roles");
+        }
+        UserEntity user = this.userRepository.findByEmail(email).orElseThrow();
+        Trip trip = this.tripRepository.findById(id).orElseThrow();
+        UserTrip userTrip = userTripRepository.findByUserAndTrip(user, trip);
+
+        TripRole tripRole = tripRoleService.getRoleByName(role);
+        userTrip.setTripRole(tripRole);
+        userTripRepository.save(userTrip);
     }
 }
